ec2: Move util to utils, and fix the imports
